refactor(db): add database module to infra folder
