+++ src/infra/database/prisma.service.ts
@@ -1,0 +1,30 @@
+import {
+  Inject,
+  Injectable,
+  type OnModuleDestroy,
+  type OnModuleInit,
+} from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
+import { PrismaClient } from 'generated/prisma';
+import { Env } from '@/core/config/env';
+
+@Injectable()
+export class PrismaService
+  extends PrismaClient
+  implements OnModuleInit, OnModuleDestroy
+{
+  constructor(@Inject(ConfigService) configService: ConfigService<Env, true>) {
+    super({
+      log:
+        configService.get('NODE_ENV', { infer: true }) === 'development'
+          ? ['query', 'error', 'warn']
+          : ['error'],
+    });
+  }
+  onModuleInit() {
+    return this.$connect();
+  }
+  onModuleDestroy() {
+    return this.$disconnect();
+  }
+}
